backend/requirements.txt
# Web Framework
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0

# Environment Variables
python-dotenv==1.0.0

# LangChain Ecosystem
langchain>=1.1.0
langchain-community>=0.4.1
langchain-chroma==0.1.2
langchain-groq>=1.1.0

# Vector Database
chromadb==0.5.5

# LLM API
groq==0.8.0

# HTTP Requests
requests==2.31.0

# PDF Processing
pypdf==3.17.4

# Data Validation
pydantic==2.8.2

# Embeddings and ML
sentence-transformers==2.2.2
torch==2.0.1

# Additional dependencies for sentence-transformers and embeddings
transformers>=4.30.0
numpy>=1.24.0
scikit-learn>=1.3.0

# HuggingFace Hub (for API embeddings)
huggingface-hub>=0.33.4
langchain-huggingface>=1.1.0
